+++ R/StartPELVIS.R
@@ -1,3 +1,0 @@
-StartPELVIS <- function() {
-  runApp(system.file("PELVISShinyApp", package="PELVIS"), launch.browser=TRUE)
-}